Do not spawn a new process on each D3D call.
Everything is obviously a lot faster with a persistent d3d4linux.exe. Of
course there is one instance per thread, so that whatever the caller had
in mind in terms of parallel processing is retained. The drawback is that
the children are never killed except at the end of the parent, but this
seems acceptable.